refactor: relocate decorators
