+++ .storybook/decorators/nexti18next.decorator.js
@@ -1,5 +1,5 @@
 import { I18nextProvider } from 'react-i18next';
-import i18n from './i18n';
+import i18n from '../i18n';
 
 export const withNexti18next = (Story, context) => {
   return (
